feat: move the recorder interface into modal
