+++ lib/beamer/locations/journal_location.dart
@@ -2,7 +2,6 @@ import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/features/journal/ui/pages/entry_details_page.dart';
 import 'package:lotti/features/journal/ui/pages/infinite_journal_page.dart';
-import 'package:lotti/features/speech/ui/pages/switchable_record_audio_page.dart';
 import 'package:lotti/features/surveys/ui/fill_survey_page.dart';
 import 'package:lotti/utils/uuid.dart';
 
@@ -13,7 +12,6 @@ class JournalLocation extends BeamLocation<BeamState> {
   List<String> get pathPatterns => [
         '/journal',
         '/journal/:entryId',
-        '/journal/:entryId/record_audio/:linkedId',
         '/journal/fill_survey/:surveyType',
       ];
 
@@ -23,8 +21,6 @@ class JournalLocation extends BeamLocation<BeamState> {
     bool pathContainsKey(String s) => state.pathParameters.containsKey(s);
 
     final entryId = state.pathParameters['entryId'];
-    final linkedId = state.pathParameters['linkedId'];
-    final categoryId = state.queryParameters['categoryId'];
 
     return [
       const BeamPage(
@@ -43,12 +39,6 @@ class JournalLocation extends BeamLocation<BeamState> {
           child: FillSurveyWithTypePage(
             surveyType: state.pathParameters['surveyType'],
           ),
-        ),
-      if (pathContains('record_audio/'))
-        BeamPage(
-          key: ValueKey('record_audio-$linkedId'),
-          child: SwitchableRecordAudioPage(
-              linkedId: linkedId, categoryId: categoryId),
         ),
     ];
   }
