+++ lib/beamer/locations/tasks_location.dart
@@ -1,7 +1,6 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/features/journal/ui/pages/infinite_journal_page.dart';
-import 'package:lotti/features/speech/ui/pages/switchable_record_audio_page.dart';
 import 'package:lotti/features/tasks/ui/pages/task_details_page.dart';
 import 'package:lotti/utils/uuid.dart';
 
@@ -12,16 +11,11 @@ class TasksLocation extends BeamLocation<BeamState> {
   List<String> get pathPatterns => [
         '/tasks',
         '/tasks/:taskId',
-        '/tasks/:taskId/record_audio/:linkedId',
       ];
 
   @override
   List<BeamPage> buildPages(BuildContext context, BeamState state) {
-    bool pathContains(String s) => state.uri.path.contains(s);
-
     final taskId = state.pathParameters['taskId'];
-    final linkedId = state.pathParameters['linkedId'];
-    final categoryId = state.queryParameters['categoryId'];
 
     return [
       const BeamPage(
@@ -33,12 +27,6 @@ class TasksLocation extends BeamLocation<BeamState> {
         BeamPage(
           key: ValueKey('tasks-$taskId'),
           child: TaskDetailsPage(taskId: taskId!),
-        ),
-      if (pathContains('record_audio/'))
-        BeamPage(
-          key: ValueKey('record_audio-$linkedId'),
-          child: SwitchableRecordAudioPage(
-              linkedId: linkedId, categoryId: categoryId),
         ),
     ];
   }
